Remove onchain pay from core
