--- src/core/lightning/index.types.d.ts
@@ -1,22 +1,0 @@
-type AuthenticatedLnd = import("lightning").AuthenticatedLnd
-
-interface IAddInvoiceRequest {
-  value: number
-  memo: string | undefined
-  selfGenerated?: boolean
-}
-
-interface IFeeRequest {
-  amount?: number
-  invoice?: string
-  username?: string
-}
-
-interface IPaymentRequest {
-  username?: string
-  amount?: number
-  invoice?: string
-  memo?: string
-  isReward?: boolean
-  twoFAToken?: string
-}
