+++ src/services/lnd/index.types.d.ts
@@ -1,0 +1,1 @@
+type AuthenticatedLnd = import("lightning").AuthenticatedLnd
